requirements.txt
albumentations==0.5.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imgaug==0.4.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
kaggle==1.5.12
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.1
Shapely==1.7.1
six==1.15.0
text-unidecode==1.3
tifffile==2021.3.5
timm==0.4.5
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
